Move 'test_no_implemented_methods' to its own test file
